漏了一步：移除无用的类型转换 > 升级 MybatisPlus 版本 2.1.9 -> 3.4.3.3 9cb7eeed

+++ candy-core/src/main/java/com/cc/api/utils/DB.java
@@ -42,7 +42,7 @@ public class DB {
         Result<List<T>> result = new Result<>(page.getRecords());
         result.curPage = page.getCurrent();
         result.pageSize = page.getSize();
-        result.totalCount = (long) page.getTotal();
+        result.totalCount = page.getTotal();
         return result;
     }
